% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_model}
\alias{build_model}
\title{Build Keras model}
\usage{
build_model(
  hidden.layer,
  input_shape,
  output_units = 1,
  output_activation = "sigmoid",
  hidden_activation = "relu",
  dropout_rate = NULL
)
}
\arguments{
\item{hidden.layer}{vector of integers for number of hidden units in each
hidden layer}

\item{input_shape}{integer for number of input features}

\item{output_units}{integer for number of output units, default is 1}

\item{output_activation}{string for output layer activation function,
default is "sigmoid"}

\item{hidden_activation}{string for hidden layer activation function,
default is "relu"}

\item{dropout_rate}{double or vector for proportion of hidden layer to drop out.}
}
\value{
Keras model object
}
\description{
Specify model for deep learning
}
\keyword{internal}
