% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{check_cran_deps}
\alias{check_cran_deps}
\title{Check for required CRAN packages and prompt installation if missing.}
\usage{
check_cran_deps()
}
\value{
Invisibly returns TRUE if all required packages are installed.
}
\description{
Check for required CRAN packages and prompt installation if missing.
}
\keyword{internal}
