% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{check_python_modules}
\alias{check_python_modules}
\title{Check for required Python modules and prompt installation if missing.}
\usage{
check_python_modules()
}
\value{
Invisibly returns TRUE if all required modules are available.
}
\description{
Check for required Python modules and prompt installation if missing.
}
\keyword{internal}
