% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{complier_mod}
\alias{complier_mod}
\title{PATT_C SL Ensemble}
\usage{
complier_mod(data, ID = NULL, SL.library = NULL, tr)
}
\arguments{
\item{data}{list object of data}

\item{tr}{}

\item{SL.library.}{Employs extreme gradient boosting, elastic net regression, random forest, and neural nets.}
}
\description{
\code{PATT_C SL Ensemble} estimates the Population Average Treatment Effect of the Treated from experimental data with noncompliers
using the super learner ensemble that includes extreme gradient boosting, glmnet (elastic net regression), random forest and neural nets.
}
