% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{complier_predict}
\alias{complier_predict}
\title{Complier model prediction}
\usage{
complier_predict(complier.mod, exp.data, treat.var, compl.var)
}
\arguments{
\item{complier.mod}{output from trained ensemble superlearner model.}

\item{exp.data}{experimental dataset}
}
\value{
\code{data.frame} object with true compliers, predicted compliers in the
control group, and all compliers (actual + predicted).
}
\description{
Predict Compliance from control group in experimental data
}
