% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL.gbm}
\alias{create.SL.gbm}
\title{create.SL.gbm}
\usage{
create.SL.gbm(distribution = c("bernoulli", "adaboost", "gaussian"))
}
\arguments{
\item{distribution}{}
}
\description{
create.SL.gbm
}
\keyword{internal}
