% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL.knn}
\alias{create.SL.knn}
\title{Title}
\usage{
create.SL.knn(k = c(20, 30, 40, 50))
}
\arguments{
\item{k}{}
}
\description{
Creates knn wrappers in the global environment with different nearest neighbors.
The default value for k in SL.knn is 10
}
