% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL.randomForest}
\alias{create.SL.randomForest}
\title{create.SL.randomForest}
\usage{
create.SL.randomForest(tune = list(mtry = c(1, 5, 10), nodesize = c(1, 5)))
}
\arguments{
\item{tune}{default value set to tune = list(mtry = c(1, 5, 10), nodesize = c(1, 5))}
}
\value{
wrapper for Random Forest learner to be exported to Global Environment
}
\description{
Creates additional randomForest wrappers.
}
\keyword{internal}
