% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL}
\alias{create.SL}
\title{create.SL}
\usage{
create.SL(learners = "all")
}
\arguments{
\item{learners}{vector of ML wrappers to be used for ensemble method}
}
\value{
vector of ML wrappers
}
\description{
create and message ML wrappers for ensemble methods.
}
\keyword{internal}
