% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{deep_predict}
\alias{deep_predict}
\title{Title}
\usage{
deep_predict(
  deep.complier.mod,
  complier.formula,
  exp.data,
  treat.var,
  compl.var
)
}
\arguments{
\item{deep.complier.mod}{}

\item{complier.formula}{}

\item{exp.data}{}

\item{treat.var}{}

\item{compl.var}{}
}
\description{
Title
}
