% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{deep_response_model}
\alias{deep_response_model}
\title{Title}
\usage{
deep_response_model(
  response.formula,
  exp.data,
  deep.compliers,
  compl.var,
  algorithm = "adam",
  hidden.layer = c(2, 2),
  hidden_activation = "relu",
  epoch = 10,
  verbose = 1,
  batch_size = 32,
  output_units = 1,
  validation_split = NULL,
  output_activation = "linear",
  loss = "mean_squared_error",
  metrics = "mean_absolute_error"
)
}
\arguments{
\item{response.formula}{}

\item{exp.data}{}

\item{deep.compliers}{}

\item{compl.var}{}

\item{algorithm}{}

\item{hidden.layer}{}

\item{hidden_activation}{}

\item{epoch}{}

\item{verbose}{}

\item{batch_size}{}

\item{output_activation}{}

\item{loss}{}

\item{metrics}{}
}
\description{
Title
}
