% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{define.SL.class.library}
\alias{define.SL.class.library}
\title{define.SL.class.library}
\usage{
define.SL.class.library(
  SL.library.class = c("SL.gbm.adaboost", "SL.gbm.bernoulli", "SL.glmnet", "SL.xgboost",
    "SL.randomForest.3")
)
}
\arguments{
\item{SL.library.class}{}
}
\description{
define.SL.class.library
}
\keyword{internal}
