% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Survey Experiment of Support for Populist Policy}
\format{
\subsection{\code{exp_data}}{

A data frame with 257 rows and 12 columns:
\describe{
\item{female}{ Gender.}
\item{age}{ Age of participant.}
\item{income}{ Monthly household income.}
\item{religion}{ Religious denomination}
\item{practicing_religion}{ Importance of religion in life.}
\item{education}{ Educational level of participant.}
\item{political_ideology}{ Political ideology of participant.}
\item{employment}{ Employment status of participant.}
\item{marital_status}{ Marital status of participant.}
\item{job_loss}{ Concern about job loss.}
\item{strong_leader}{ Binary treatment measure of leader type.}
\item{support_war}{ Binary outcome measure for willingness to fight war.}
#'   ...
}
}
}
\source{
Yadav and Mukherjee (2024)
}
\usage{
data(exp_data)
}
\description{
Shortened version of survey response data that incorporates a vignette survey
experiment. The vignette describes an international crisis between country A
and B. After reading this vignette, respondents are randomly assigned to the
control group or to one of two treatments: policy prescription to said crisis
by strong (populist) leader and centrist (non-populist) leader. The
respondents are then asked whether they are willing to support the policy
decision to fight a war against country A, which is the dependent variable.
}
\keyword{dataset}
