% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_data_full}
\alias{exp_data_full}
\title{Survey Experiment of Support for Populist Policy}
\format{
\subsection{\code{exp_data_full}}{

A data frame with 514 rows and 12 columns:
\describe{
\item{female}{ Gender.}
\item{age}{ Age of participant.}
\item{income}{ Monthly household income.}
\item{religion}{ Religious denomination}
\item{practicing_religion}{ Importance of religion in life.}
\item{education}{ Educational level of participant.}
\item{political_ideology}{ Political ideology of participant.}
\item{employment}{ Employment status of participant.}
\item{marital_status}{ Marital status of participant.}
\item{job_loss}{ Concern about job loss.}
\item{strong_leader}{ Binary treatment measure of leader type.}
\item{support_war}{ Binary outcome measure for willingness to fight war.}
#'   ...
}
}
}
\source{
Yadav and Mukherjee (2024)
}
\usage{
data(exp_data_full)
}
\description{
Extended experiment data with 514 observations
}
\keyword{dataset}
