% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expcall}
\alias{expcall}
\title{create list object of experimental data for easy processing}
\usage{
expcall(
  response.formula,
  treat.var,
  compl.var,
  exp.data,
  weights = NULL,
  cluster = NULL,
  ID = NULL
)
}
\arguments{
\item{ID}{}
}
\description{
create list object of experimental data for easy processing
}
\keyword{internal}
