% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{meta_learner}
\alias{meta_learner}
\title{S_T-learner Ensemble}
\usage{
meta_learner(
  data,
  cov.formula,
  treat.var,
  meta.learner.type,
  learners = c("SL.glmnet", "SL.xgboost", "SL.ranger", "SL.nnet"),
  nfolds = 5,
  seed = 1234
)
}
\arguments{
\item{data}{list object of data}

\item{seed}{}

\item{meta.learner.type.}{This is the S-learner and T-learner model.}

\item{learners.}{The super learner ensemble that includes extreme gradient boosting, glmnet, random forest, and neural nets.}
}
\description{
\code{S_T-learner Ensemble} implements the S-learner and T-learner for estimating CATE using the super learner ensemble method.
The super learner in this case includes the following machine learning algorithms: extreme gradient boosting, glmnet (elastic net regression),
random forest and neural nets.
}
