% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{Meta_learner_ensemble}
\alias{Meta_learner_ensemble}
\title{Meta_learner_ensemble}
\usage{
Meta_learner_ensemble(
  data,
  cov.formula,
  treat.var,
  meta.learner.type,
  learners = c("SL.glmnet", "SL.xgboost", "SL.ranger", "SL.nnet"),
  nfolds = 5
)
}
\arguments{
\item{data}{\code{data.frame} object of data}

\item{cov.formula}{formula description of the model y ~ x(list of covariates)}

\item{treat.var}{string for the name of treatment variable.}

\item{meta.learner.type}{string specifying is the S-learner and
\code{"T.Learner"} for the T-learner model.}

\item{learners}{vector for super learner ensemble that includes extreme gradient
boosting, glmnet, random forest, and neural nets.}

\item{nfolds}{number of folds for cross-validation. Currently supports up to
5 folds.}
}
\value{
vector of CATEs estimated by the meta learners for each observation.
}
\description{
\code{Meta_learner_ensemble} implements the S-learner and T-learner for
estimating CATE using the super learner ensemble method. The super learner in
this case includes the following machine learning algorithms:
extreme gradient boosting, glmnet (elastic net regression), random forest and
neural nets.
}
\examples{
# load dataset
data(exp_data)
# estimate CATEs with S Learner
control <- SuperLearner::SuperLearner.CV.control(V=5)
# estimate CATEs with S Learner
slearner <- Meta_learner_ensemble(cov.formula = support_war ~ age +
                                  income + employed + job_loss,
                                data = exp_data,
                                treat.var = "strong_leader",
                                meta.learner.type = "S.Learner",
                                learners = c("SL.glmnet","SL.xgboost"),
                                nfolds = 5)
# estimate CATEs with T Learner
\donttest{
tlearner <- Meta_learner_ensemble(cov.formula = support_war ~ age + income +
                                  employed  + job_loss,
                                  data = exp_data,
                                  treat.var = "strong_leader",
                                  meta.learner.type = "T.Learner",
                                  learners = c("SL.glmnet","SL.xgboost"),
                                  nfolds = 5)
                                  }
\dontrun{
tlearner <- Meta_learner_ensemble(cov.formula = support_war ~ age + income  +
                                                employed  + job_loss,
                                  data = exp_data,
                                  treat.var = "strong_leader",
                                  meta.learner.type = "R.Learner",
                                  learners = c("SL.glmnet","SL.xgboost"),
                                  nfolds = 5)
                                  }

}
