% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_meta_learners.R
\name{metalearner_deep}
\alias{metalearner_deep}
\title{Title}
\usage{
metalearner_deep(
  data,
  cov.formula,
  treat.var,
  meta.learner.type,
  nfolds = 5,
  algorithm = "adam",
  hidden.layer = c(2, 2),
  hidden_activation = "relu",
  output_activation = "linear",
  output_units = 1,
  loss = "mean_squared_error",
  metrics = "mean_absolute_error",
  epoch = 10,
  verbose = 1,
  batch_size = 32,
  validation_split = NULL
)
}
\arguments{
\item{data}{}

\item{cov.formula}{}

\item{treat.var}{}

\item{meta.learner.type}{}

\item{nfolds}{}

\item{algorithm}{}

\item{hidden.layer}{}

\item{hidden_activation}{}

\item{output_activation}{}

\item{output_units}{}

\item{loss}{}

\item{metrics}{}

\item{epoch}{}

\item{verbose}{}

\item{batch_size}{}

\item{validation_split}{}
}
\description{
Title
}
