% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{neuralnet_complier_mod}
\alias{neuralnet_complier_mod}
\title{PATT_C DeepNN}
\usage{
neuralnet_complier_mod(
  complier.formula,
  exp.data,
  treat.var,
  algorithm = "rprop+",
  hidden.layer = c(4, 2),
  ID = NULL,
  stepmax = 1e+05
)
}
\arguments{
\item{stepmax}{}

\item{hidden.layer.}{The Resilient back propagation (Rprop) algorithm for training neural networks.}
}
\description{
\code{PATT_C DeepNN} estimates the Population Average Treatment Effect of the Treated from experimental data with noncompliers
using Deep Neural Networks. The Resilient back propagation (Rprop) algorithm is used for training neural networks.
}
