% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{neuralnet_pattc_counterfactuals}
\alias{neuralnet_pattc_counterfactuals}
\title{Assess Population Data counterfactuals}
\usage{
neuralnet_pattc_counterfactuals(
  pop.data,
  neuralnet.response.mod,
  ID = NULL,
  cluster = NULL
)
}
\arguments{
\item{cluster}{}
}
\description{
Assess Population Data counterfactuals
}
