% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{neuralnet_predict}
\alias{neuralnet_predict}
\title{Predicting Compliance from experimental data}
\usage{
neuralnet_predict(neuralnet.complier.mod, exp.data, treat.var, compl.var)
}
\arguments{
\item{compl.var}{}
}
\description{
Predicting Compliance from experimental data
}
