% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{neuralnet_response_model}
\alias{neuralnet_response_model}
\title{Modeling Responses from experimental data Using Deep NN}
\usage{
neuralnet_response_model(
  response.formula,
  exp.data,
  neuralnet.compliers,
  compl.var,
  algorithm = "rprop+",
  hidden.layer = c(4, 2),
  stepmax = 1e+05
)
}
\arguments{
\item{stepmax}{}
}
\description{
Modeling Responses from experimental data Using Deep NN
}
