% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{patt_ensemble}
\alias{patt_ensemble}
\title{Estimate PATT_C using SL ensemble}
\usage{
patt_ensemble(
  response.formula,
  exp.data,
  pop.data,
  treat.var,
  compl.var,
  createSL = TRUE,
  ID = NULL,
  cluster = NULL,
  bootse = FALSE,
  bootp = FALSE,
  bootn = 999,
  samedata = FALSE,
  equivalence = FALSE
)
}
\arguments{
\item{equivalence}{}
}
\description{
Estimate PATT_C using SL ensemble
}
