% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{patt_neural}
\alias{patt_neural}
\title{Estimate PATT_C using Deep NN}
\usage{
patt_neural(
  response.formula,
  exp.data,
  pop.data,
  treat.var,
  compl.var,
  compl.algorithm = "rprop+",
  response.algorithm = "rprop+",
  compl.hidden.layer = c(4, 2),
  response.hidden.layer = c(4, 2),
  compl.stepmax = 1e+05,
  response.stepmax = 1e+05,
  ID = NULL,
  cluster = NULL,
  bootse = FALSE,
  bootp = FALSE,
  bootn = 999,
  samedata = FALSE,
  equivalence = FALSE
)
}
\arguments{
\item{equivalence}{}
}
\description{
Estimate PATT_C using Deep NN
}
