% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{pattc_counterfactuals}
\alias{pattc_counterfactuals}
\title{Assess Population Data counterfactuals}
\usage{
pattc_counterfactuals(
  pop.data,
  response.mod,
  id = NULL,
  cluster = NULL,
  cut.point = 0.5
)
}
\arguments{
\item{cut.point}{}
}
\description{
Assess Population Data counterfactuals
}
