% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{pattc_deep}
\alias{pattc_deep}
\title{Deep PATT-C}
\usage{
pattc_deep(
  response.formula,
  compl.var,
  treat.var,
  exp.data,
  pop.data,
  compl.algorithm = "adam",
  response.algorithm = "adam",
  compl.hidden.layer = c(4, 2),
  response.hidden.layer = c(4, 2),
  compl.hidden_activation = "relu",
  response.hidden_activation = "relu",
  response.output_activation = "linear",
  response.output_units = 1,
  response.loss = "mean_squared_error",
  response.metrics = "mean_absolute_error",
  ID = NULL,
  weights = NULL,
  cluster = NULL,
  compl.epoch = 10,
  response.epoch = 10,
  compl.validation_split = NULL,
  response.validation_split = NULL,
  verbose = 1,
  batch_size = 32,
  binary.preds = FALSE,
  bootstrap = FALSE,
  nboot = 1000
)
}
\arguments{
\item{response.formula}{A formula specifying the response variable and covariates.}

\item{compl.var}{A string specifying the name of the compliance variable.}

\item{treat.var}{A string specifying the name of the treatment variable.}

\item{exp.data}{A data frame containing the experimental data.}

\item{pop.data}{A data frame containing the population data.}

\item{compl.algorithm}{}

\item{response.algorithm}{}

\item{compl.hidden.layer}{}

\item{response.hidden.layer}{}

\item{compl.hidden_activation}{}

\item{response.hidden_activation}{}

\item{response.output_activation}{}

\item{response.output_units}{}

\item{response.loss}{}

\item{response.metrics}{}

\item{ID}{An optional string specifying the name of the identifier variable.}

\item{weights}{An optional string specifying the name of the weights variable.}

\item{cluster}{An optional string specifying the name of the clustering variable.}

\item{compl.epoch}{}

\item{response.epoch}{Integer for the number of epochs for response model.}

\item{compl.validation_split}{}

\item{response.validation_split}{}

\item{verbose}{An integer specifying the verbosity level during training. Default is 1.}

\item{batch_size}{An integer specifying the batch size for training the deep learning models. Default is 32.}

\item{binary.preds}{A logical indicating whether to treat predictions as binary outcomes. Default is FALSE.}

\item{bootstrap}{A logical indicating whether to use bootstrapping for confidence intervals. Default is FALSE.}

\item{nboot}{An integer specifying the number of bootstrap samples if bootstrap is TRUE. Default is 1000.}
}
\value{
pattc_deep object containing the fitted models, predictions, counterfactuals, and PATT-C estimate.
}
\description{
This function implements the Deep PATT-C method for estimating the Population Average Treatment
Effect on the Treated Compliers (PATT-C) using deep learning models.
It consists of training a deep learning model to predict compliance among treated individuals,
predicting compliance in the experimental data, training a response model among predicted compliers,
and estimating counterfactual outcomes in the population data.
}
\examples{
\dontrun{
library(DeepLearningCausal)
library(magrittr)
data("exp_data")
data("pop_data")
set.seed(1243)
deeppattc <- pattc_deep(response.formula = support_war ~ age + female +
income + education +  employed + married +
  hindu + job_loss,
exp.data = exp_data,
pop.data = pop_data,
treat.var = "strong_leader",
compl.var = "compliance",
algorithm = "adam",
hidden.layer = c(2,2),
ID = NULL,
weights = NULL,
cluster = NULL,
epoch = 50,
verbose = 1,
batch_size = 32,
model_type = "classification",
binary.preds = FALSE,
boot = FALSE
)
}
}
