% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{pattc_deep_counterfactuals}
\alias{pattc_deep_counterfactuals}
\title{Title}
\usage{
pattc_deep_counterfactuals(
  pop.data,
  response.mod,
  response_formula,
  ID = NULL,
  cluster = NULL,
  binary.preds = FALSE
)
}
\arguments{
\item{pop.data}{}

\item{response.mod}{}

\item{response_formula}{}

\item{ID}{}

\item{cluster}{}

\item{binary.preds}{}
}
\description{
Title
}
