% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.metalearner_neural}
\alias{plot.metalearner_neural}
\title{plot.metalearner_neural}
\usage{
\method{plot}{metalearner_neural}(x, ..., conf_level = 0.95, type = "CATEs")
}
\arguments{
\item{x}{\code{metalearner_neural} model object.}

\item{...}{Additional arguments}

\item{conf_level}{numeric value for confidence level. Defaults to 0.95.}

\item{type}{"CATEs" or "predict".}
}
\value{
\code{ggplot} object.
}
\description{
Uses \code{plot()} to generate histogram of distribution of CATEs or predicted
outcomes from  \code{metalearner_neural}
}
