% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pattc_deeplearning}
\alias{plot.pattc_deeplearning}
\title{plot.pattc_deeplearning}
\usage{
\method{plot}{pattc_deeplearning}(x, ...)
}
\arguments{
\item{x}{\code{pattc_deeplearning} model object}

\item{...}{Additional arguments}
}
\value{
\code{ggplot} object
}
\description{
Uses \code{plot()} to generate histogram of distribution of predicted
outcomes from  \code{pattc_deeplearning}
}
