% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pattc_ensemble}
\alias{plot.pattc_ensemble}
\title{plot.pattc_ensemble}
\usage{
\method{plot}{pattc_ensemble}(model_obj)
}
\arguments{
\item{model_obj}{\code{pattc_ensemble} model object}
}
\value{
\code{ggplot} object
}
\description{
Uses \code{plot()} to generate histogram of ditribution of CATEs or predicted
outcomes from  \code{pattc_ensemble}
}
