% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pattc_neural}
\alias{plot.pattc_neural}
\title{plot.pattc_neural}
\usage{
\method{plot}{pattc_neural}(x, ...)
}
\arguments{
\item{x}{\code{pattc_neural} model object}

\item{...}{Additional arguments}
}
\value{
\code{ggplot} object
}
\description{
Uses \code{plot()} to generate histogram of distribution of CATEs or predicted
outcomes from  \code{pattc_neural}
}
