% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{World Value Survey India Sample}
\format{
\subsection{\code{pop_data}}{

A data frame with 846 rows and 13 columns:
\describe{
\item{female}{ Respondent’s Sex.}
\item{age}{ Age of respondent.}
\item{income}{ income group of Household.}
\item{religion}{ Religious denomination}
\item{practicing_religion}{ Importance of religion in respondent’s life.}
\item{education}{ Educational level of respondent.}
\item{political_ideology}{ Political ideology of respondent.}
\item{employment}{ Employment status and full-time employee.}
\item{marital_status}{ Marital status of respondent.}
\item{job_loss}{ Concern about job loss.}
\item{support_war}{ Binary (Yes/No) outcome measure for willingness to fight war.}
\item{strong_leader}{ Binary measure of preference for strong leader.}
...
}
}
}
\source{
Haerpfer, C., Inglehart, R., Moreno, A., Welzel, C., Kizilova, K., Diez-Medrano J., M. Lagos, P. Norris, E. Ponarin & B. Puranen et al. (eds.). 2020. World Values Survey: Round Seven – Country-Pooled Datafile. Madrid, Spain & Vienna, Austria: JD Systems Institute & WVSA Secretariat. <doi.org/10.14281/18241.1>
}
\usage{
data(pop_data)
}
\description{
World Value Survey (WVS) Data for India in 2022. The variables drawn from the
said WVS India data match the covariates from the India survey experiment sample.
}
\keyword{dataset}
