% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{popcall}
\alias{popcall}
\title{create list object of population data for easy data processing}
\usage{
popcall(
  response.formula,
  compl.var,
  pop.data,
  weights = NULL,
  cluster = NULL,
  ID = NULL
)
}
\arguments{
\item{response.formula}{formula for response equation of binary outcome
variable and covariates}

\item{compl.var}{string for complier variable}

\item{pop.data}{\code{data.frame} of experimental variable}

\item{weights}{observation weights}

\item{cluster}{clustering variable}

\item{ID}{identifier variable}
}
\value{
list of processed dataset
}
\description{
create list object of population data for easy data processing
}
\keyword{internal}
