% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{predict.SL.mean}
\alias{predict.SL.mean}
\title{predict.SL.mean}
\usage{
\method{predict}{SL.mean}(object, newdata, family, X = NULL, Y = NULL, ...)
}
\arguments{
\item{object}{\code{SL.mean} class object}

\item{newdata}{new dataset}

\item{family}{regression or classifier}

\item{X}{predictor variable}

\item{Y}{outcome variable}

\item{...}{unused}
}
\value{
predictions of outcome variable
}
\description{
predict results from \code{SL.mean}. Currently unused.
}
\keyword{internal}
