% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_meta_learners.R
\name{print.metalearner_deeplearning}
\alias{print.metalearner_deeplearning}
\title{print.metalearner_deeplearning}
\usage{
\method{print}{metalearner_deeplearning}(x, ...)
}
\arguments{
\item{x}{\code{metalearner_deeplearning} class object from \code{metalearner_deeplearning}}

\item{...}{additional parameter}
}
\value{
list of model results
}
\description{
Print method for \code{metalearner_deeplearning}
}
