% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_meta_learners.R
\name{print.metalearner_deepneural}
\alias{print.metalearner_deepneural}
\title{print.metalearner_deepneural}
\usage{
\method{print}{metalearner_deepneural}(x, ...)
}
\arguments{
\item{x}{\code{metalearner_deepneural} class object from \code{metalearner_deepneural}}

\item{...}{additional parameter}
}
\value{
list of model results
}
\description{
Print method for \code{metalearner_deepneural}
}
