% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R
\name{print.metalearner_ensemble}
\alias{print.metalearner_ensemble}
\title{print.metalearner_ensemble}
\usage{
\method{print}{metalearner_ensemble}(x, ...)
}
\arguments{
\item{x}{\code{metalearner_ensemble} class object from \code{metalearner_ensemble}}

\item{...}{additional parameter}
}
\value{
list of model results
}
\description{
Print method for \code{metalearner_ensemble}
}
