% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_meta_learners.R
\name{print.metalearner_neural}
\alias{print.metalearner_neural}
\title{print.metalearner_neural}
\usage{
\method{print}{metalearner_neural}(x, ...)
}
\arguments{
\item{x}{\code{metalearner_neural} class object from \code{metalearner_neural}}

\item{...}{additional parameter}
}
\value{
list of model results
}
\description{
Print method for \code{metalearner_neural}
}
