% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{print.pattc_deep}
\alias{print.pattc_deep}
\title{print.pattc_deep}
\usage{
\method{print}{pattc_deep}(x, ...)
}
\arguments{
\item{x}{\code{pattc_deep} class object from \code{pattc_deep}}

\item{...}{additional arguments}
}
\value{
list of model results
}
\description{
Print method for \code{pattc_deep}
}
