% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{print.pattc_deeplearning}
\alias{print.pattc_deeplearning}
\title{print.pattc_deeplearning}
\usage{
\method{print}{pattc_deeplearning}(x, ...)
}
\arguments{
\item{x}{\code{pattc_deeplearning} class object from \code{pattc_deeplearning}}

\item{...}{additional arguments}
}
\value{
list of model results
}
\description{
Print method for \code{pattc_deeplearning}
}
