% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{print.pattc_deepneural}
\alias{print.pattc_deepneural}
\title{print.pattc_deepneural}
\usage{
\method{print}{pattc_deepneural}(x, ...)
}
\arguments{
\item{x}{\code{pattc_deepneural} class object from \code{pattc_deepneural}}

\item{...}{additional parameter}
}
\value{
list of model results
}
\description{
Print method for \code{pattc_deepneural}
}
