% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{print.pattc_ensemble}
\alias{print.pattc_ensemble}
\title{print.pattc_ensemble}
\usage{
\method{print}{pattc_ensemble}(model, ...)
}
\arguments{
\item{model}{\code{pattc_ensemble} class object from \code{pattc_ensemble}}

\item{...}{additional parameter}
}
\description{
Print method for \code{pattc_ensemble}
}
