% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{print.pattc_neural}
\alias{print.pattc_neural}
\title{print.pattc_neural}
\usage{
\method{print}{pattc_neural}(x, ...)
}
\arguments{
\item{x}{\code{pattc_neural} class object from \code{pattc_neural}}

\item{...}{additional parameter}
}
\value{
list of model results
}
\description{
Print method for \code{pattc_neural}
}
