% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{response_model}
\alias{response_model}
\title{Model Responses from experimental data using SL ensemble}
\usage{
response_model(
  exp.data,
  exp.compliers,
  family = "binomial",
  ID = NULL,
  SL.library = NULL
)
}
\arguments{
\item{SL.library}{}
}
\description{
Model Responses from experimental data using SL ensemble
}
