% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_meta_learners.R
\name{ST_learner_DeepNN}
\alias{ST_learner_DeepNN}
\title{ST_learner_DeepNN}
\usage{
ST_learner_DeepNN(
  data,
  cov.formula,
  treat.var,
  meta.learner.type,
  stepmax = 1e+05,
  nfolds = 5,
  algorithm = "rprop+",
  hidden.layer = c(4, 2),
  seed = 1234,
  linear.output = FALSE
)
}
\arguments{
\item{data}{list object of data}

\item{cov.formula}{formula object for}

\item{nfolds}{number of folds for cross-validation. Currently supports up to
5 folds.}

\item{algorithm}{a string for the algorithm for the neural network.
Default set to \verb{rprop+}, the Resilient back propagation (Rprop) with weight
backtracking algorithm for training neural networks.}

\item{hidden.layer}{vector of}

\item{linear.output}{}

\item{meta.learner.type.}{This is the S-learner and T-learner model.}
}
\description{
\code{ST_learner_DeepNN} implements the S-learner and T-learner for estimating
CATE using Deep Neural Networks. The Resilient back propagation (Rprop)
algorithm is used for training neural networks.
}
