% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtc.R
\name{WtC}
\alias{WtC}
\title{WtC}
\usage{
WtC(
  x,
  y,
  c = NULL,
  weight = NULL,
  weighty = NULL,
  weightc = NULL,
  cluster = NULL,
  clustery = NULL,
  clusterc = NULL,
  samedata = TRUE,
  alternative = "two.tailed",
  mean1 = TRUE,
  bootse = TRUE,
  bootp = FALSE,
  bootn = 1000,
  drops = "pairwise",
  equivalence = FALSE
)
}
\arguments{
\item{x}{vector for group 1}

\item{y}{vector for group 2}

\item{c}{vector for compliance variable}

\item{weight}{vector for weights of x}

\item{weighty}{vector for weights of y}

\item{weightc}{vector for weights of c}

\item{cluster}{vector of clusters for x}

\item{clustery}{vector of clusters for y}

\item{clusterc}{vector of clusters for c}

\item{samedata}{logical for whether x and y are from same dataset}

\item{alternative}{string of "two.tailed", "greater", or "less"}

\item{mean1}{logical to force weights to have average of 1}

\item{bootse}{logical for bootstrapped standard errors}

\item{bootp}{logical for bootstrapped p values}

\item{bootn}{logical for number of bootstraps}

\item{drops}{string for cases of missing data}

\item{equivalence}{logical for equivalence test}
}
\value{
results of weight t test for PATT
}
\description{
Weighted t-test to estimate PATT for numeric outcome variables. For further
details ses: weights::wtd.t.test()
}
\keyword{internal}
