% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{complier_predict}
\alias{complier_predict}
\title{Predict Compliance from experimental data}
\usage{
complier_predict(complier.mod, exp.data)
}
\arguments{
\item{exp.data}{}
}
\description{
Predict Compliance from experimental data
}
