% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL.gbm}
\alias{create.SL.gbm}
\title{create.SL.gbm}
\usage{
create.SL.gbm(distribution = c("bernoulli", "adaboost", "gaussian"))
}
\arguments{
\item{distribution}{vector of distribution types, set to
c("bernoulli","adaboost","gaussian") default.}
}
\value{
wrapper for gbm learner to be exported to Global Environment
}
\description{
Creates gbm wrappers in the global environment.
}
\keyword{internal}
