% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL.randomForest}
\alias{create.SL.randomForest}
\title{Title}
\usage{
create.SL.randomForest(tune = list(mtry = c(1, 5, 10), nodesize = c(1, 5)))
}
\arguments{
\item{tune}{}
}
\description{
Title
}
