% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{deep_complier_mod}
\alias{deep_complier_mod}
\title{Title}
\usage{
deep_complier_mod(
  complier.formula,
  exp.data,
  treat.var,
  algorithm = "adam",
  hidden.layer = c(2, 2),
  hidden_activation = "relu",
  ID = NULL,
  epoch = 10,
  verbose = 1,
  batch_size = 32,
  validation_split = NULL
)
}
\arguments{
\item{complier.formula}{}

\item{exp.data}{}

\item{treat.var}{}

\item{algorithm}{}

\item{hidden.layer}{}

\item{hidden_activation}{}

\item{ID}{}

\item{epoch}{}

\item{verbose}{}

\item{batch_size}{}
}
\description{
Title
}
