% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{deep_predict}
\alias{deep_predict}
\title{Complier model prediction}
\usage{
deep_predict(
  deep.complier.mod,
  complier.formula,
  exp.data,
  treat.var,
  compl.var
)
}
\arguments{
\item{deep.complier.mod}{model object from \code{deep.complier.mod()}}

\item{complier.formula}{formula to fit compliance model (c ~ x) using
complier variable and covariates}

\item{exp.data}{\code{data.frame} object of experimental dataset}

\item{treat.var}{string specifying the binary treatment variable}

\item{compl.var}{string specifying binary complier variable}
}
\value{
\code{data.frame} object with true compliers, predicted compliers in the
control group, and all compliers (actual + predicted).
}
\description{
Predict Compliance from control group in experimental data
}
