% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IND_exp_data}
\alias{IND_exp_data}
\title{Survey Experiment of Support for Populist Policy}
\format{
\subsection{\code{IND_exp_data}}{

A data frame with 257 rows and 12 columns:
\describe{
\item{Female}{ Gender.}
\item{Age}{ Age of participant.}
\item{Income}{ Monthly household income.}
\item{Religion}{ Religious denomination}
\item{Imp_rel}{ Importance of religion in life.}
\item{Education}{ Educational level of participant.}
\item{Ideol_lr}{ Political ideology of participant.}
\item{Empl_status}{ Employment status of participant.}
\item{Marital_status}{ Marital status of participant.}
\item{job_worry}{ Concern about job loss.}
\item{Exp1trt}{ Binary treatment measure of leader type.}
\item{Exp1_dv1}{ Binary outcome measure for willingness to fight war.}
#'   ...
}
}
}
\source{
Yadav and Mukherjee (2024)
}
\usage{
data(IND_exp_data)
}
\description{
Shortened version of survey response data that incorporates a vignette survey
experiment. The vignette describes an international crisis between country A
and B. After reading this vignette, respondents are randomly assigned to the
control group or to one of two treatments: policy prescription to said crisis
by strong (populist) leader and centrist (non-populist) leader. The
respondents are then asked whether they are willing to support the policy
decision to fight a war against country A, which is the dependent variable.
}
\keyword{dataset}
