% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IND_pop_data}
\alias{IND_pop_data}
\title{World Value Survey India Sample}
\format{
\subsection{\code{IND_pop_data}}{

A data frame with 11,813 rows and 13 columns:
\describe{
\item{Female}{ Respondent’s Sex.}
\item{Age}{ Age of respondent.}
\item{Income}{ Income group of Household.}
\item{Religion}{ Religious denomination}
\item{Imp_rel}{ Importance of religion in respondent’s life.}
\item{Education}{ Educational level of respondent.}
\item{Ideol}{ Political ideology of respondent.}
\item{Empl_status}{ Employment status and full-time employee.}
\item{Marital}{ Marital status of respondent.}
\item{job_worry}{ Concern about job loss.}
\item{Exp1_trt}{ Binary treatment measure of leader type.}
\item{Exp1_dv_willing}{ Binary (Yes/No) outcome measure for willingness to fight war.}
\item{strong_leader}{ Binary measure of preference for strong leader.}
...
}
}
}
\source{
Haerpfer, C., Inglehart, R., Moreno, A., Welzel, C., Kizilova, K., Diez-Medrano J., M. Lagos, P. Norris, E. Ponarin & B. Puranen et al. (eds.). 2020. World Values Survey: Round Seven – Country-Pooled Datafile. Madrid, Spain & Vienna, Austria: JD Systems Institute & WVSA Secretariat. <doi.org/10.14281/18241.1>
}
\usage{
data(IND_pop_data)
}
\description{
World Value Survey (WVS) Data for India for the following years in which the
WVS implemented their survey questionnaire in the country: 1995, 2001, 2006,
2012, and 2022. The variables drawn from the said WVS India data match the
covariates from the India survey experiment sample.
}
\keyword{dataset}
