% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{patt_ensemble}
\alias{patt_ensemble}
\title{PATT_C SL Ensemble}
\usage{
patt_ensemble(
  response.formula,
  exp.data,
  pop.data,
  treat.var,
  compl.var,
  createSL = TRUE,
  SL.library = NULL,
  ID = NULL,
  cluster = NULL,
  bootse = FALSE,
  bootp = FALSE,
  bootn = 999
)
}
\arguments{
\item{response.formula}{formula for the effects of covariates on outcome
variable (y ~ x).}

\item{exp.data}{\code{data.frame} object for experimental data. Must include
binary treatment and compliance variable.}

\item{pop.data}{\code{data.frame} object for population data. Must include
binary compliance variable.}

\item{treat.var}{string for binary treatment variable.}

\item{compl.var}{string for binary compliance variable.}

\item{createSL}{logical. If \code{TRUE} will call on \code{create.SL} to create
SL wrappers.}

\item{SL.library}{}

\item{ID}{string for name of identifier.}

\item{cluster}{string for name of cluster variable.}

\item{bootse}{logical for bootstrapped standard errors.}

\item{bootp}{logical for bootstrapped p values.}

\item{bootn}{number of bootstrap sample.}
}
\value{
results of weighted t test as PATTC estimate.
}
\description{
\code{PATT_C_SL_Ensemble} estimates the Population Average Treatment Effect
of the Treated from experimental data with noncompliers
using the super learner ensemble that includes extreme gradient boosting,
glmnet (elastic net regression), random forest and neural nets.
}
\examples{
\donttest{
# load datasets
data(exp_data) #experimental data
data(pop_data) #population data
#attach SuperLearner (model will not recognize learner if package is not loaded)
library(SuperLearner)
#specify models and estimate PATTC
pattc_ensemble <- patt_ensemble(response.formula = support_war ~ age + income +
                                education + employed + job_loss,
                                exp.data = exp_data,
                                pop.data = pop_data,
                                treat.var = "strong_leader",
                                compl.var = "compliance",
                                createSL = TRUE,
                                SL.library = c("SL.gbm.adaboost",
                                               "SL.gbm.bernoulli",
                                               "SL.glmnet"),
                                ID = NULL,
                                cluster = NULL,
                                bootse = FALSE,
                                bootp = FALSE,
                                bootn = 999)
summary(pattc)
}
}
