% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{pattc_counterfactuals}
\alias{pattc_counterfactuals}
\title{Assess Population Data counterfactuals}
\usage{
pattc_counterfactuals(
  pop.data,
  response.mod,
  ID = NULL,
  cluster = NULL,
  potential.outcome = TRUE
)
}
\arguments{
\item{pop.data}{population dataset}

\item{response.mod}{trained model from \code{response_model}.}

\item{potential.outcome}{}
}
\description{
Create counterfactual datasets in the population for compliers and
noncompliers. Then predict potential outcomes from counterfactuals.
}
