% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{pattc_counterfactuals}
\alias{pattc_counterfactuals}
\title{Assess Population Data counterfactuals}
\usage{
pattc_counterfactuals(
  pop.data,
  response.mod,
  ID = NULL,
  cluster = NULL,
  binary.outcome = TRUE
)
}
\arguments{
\item{pop.data}{population dataset}

\item{response.mod}{trained model from \code{response_model}.}

\item{ID}{string fir identifier variable}

\item{cluster}{string for clustering variable}

\item{binary.outcome}{logical specifying whether predicted outcomes are
proportions or binary (0-1).}
}
\value{
\code{data.frame} object of predicted outcomes of counterfactual groups.
}
\description{
Create counterfactual datasets in the population for compliers and
noncompliers. Then predict potential outcomes from counterfactuals.
}
