% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_deep.R
\name{pattc_deep_counterfactuals}
\alias{pattc_deep_counterfactuals}
\title{Assess Population Data counterfactuals}
\usage{
pattc_deep_counterfactuals(
  pop.data,
  response.mod,
  response.formula,
  ID = NULL,
  cluster = NULL,
  binary.preds = FALSE
)
}
\arguments{
\item{pop.data}{population dataset}

\item{response.mod}{trained model from \code{response_model}.}

\item{response.formula}{formula specifying the response variable and covariates.}

\item{ID}{string fir identifier variable}

\item{cluster}{string for clustering variable}

\item{binary.preds}{logical for predictions to be binary or proportions when}
}
\value{
\code{data.frame} object of predicted outcomes of counterfactual groups.
}
\description{
Create counterfactual datasets in the population for compliers and
noncompliers.
}
