% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.metalearner_deepneural}
\alias{plot.metalearner_deepneural}
\title{plot.metalearner_deepneural}
\usage{
\method{plot}{metalearner_deepneural}(model_obj, type = "CATEs")
}
\arguments{
\item{model_obj}{\code{metalearner_deepneural} model object.}

\item{type}{"CATEs" or "predict".}
}
\value{
\code{ggplot} object.
}
\description{
Uses \code{plot()} to generate histogram of ditribution of CATEs or predicted
outcomes from  \code{metalearner_deepneural}
}
