% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.metalearner_ensemble}
\alias{plot.metalearner_ensemble}
\title{plot.metalearner_ensemble}
\usage{
\method{plot}{metalearner_ensemble}(model_obj, conf_level = 0.95, type = "CATEs")
}
\arguments{
\item{model_obj}{\code{metalearner_ensemble} model object}

\item{conf_level}{numeric value for confidence level. Defaults to 0.95.}

\item{type}{"CATEs" or "predict"}
}
\value{
\code{ggplot} object
}
\description{
Uses \code{plot()} to generate histogram of ditribution of CATEs or predicted
outcomes from  \code{metalearner_ensemble}
}
