% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pattc_deepneural}
\alias{plot.pattc_deepneural}
\title{plot.pattc_deepneural}
\usage{
\method{plot}{pattc_deepneural}(model_obj)
}
\arguments{
\item{model_obj}{\code{pattc_deepneural} model object}
}
\value{
\code{ggplot} object
}
\description{
Uses \code{plot()} to generate histogram of ditribution of CATEs or predicted
outcomes from  \code{pattc_deepneural}
}
