% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pattc_ensemble}
\alias{plot.pattc_ensemble}
\title{plot.pattc_ensemble}
\usage{
\method{plot}{pattc_ensemble}(x, ...)

\method{plot}{pattc_ensemble}(x, ...)
}
\arguments{
\item{x}{\code{pattc_deeplearning} model object}

\item{...}{Additional arguments}
}
\value{
\code{ggplot} object

\code{ggplot} object
}
\description{
Uses \code{plot()} to generate histogram of ditribution of CATEs or predicted
outcomes from  \code{pattc_ensemble}

Uses \code{plot()} to generate histogram of ditribution of CATEs or predicted
outcomes from  \code{pattc_deeplearning}
}
