% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{popcall}
\alias{popcall}
\title{Create list for population data}
\usage{
popcall(
  response.formula,
  compl.var,
  treat.var,
  pop.data,
  weights = NULL,
  cluster = NULL,
  ID = NULL,
  patt = TRUE
)
}
\arguments{
\item{response.formula}{formula for response equation of binary outcome
variable and covariates}

\item{compl.var}{string for complier variable}

\item{treat.var}{string for treatment variable}

\item{pop.data}{\code{data.frame} of experimental variable}

\item{weights}{observation weights}

\item{cluster}{clustering variable}

\item{ID}{identifier variable}

\item{patt}{logical for patt, subsetting population treated observations}
}
\value{
list of processed dataset
}
\description{
create list object of population data for easy data processing
}
\keyword{internal}
