% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learners.R, R/neural_meta_learners.R,
%   R/patt_ensemble.R, R/patt_neural.R
\name{print.pattc_ensemble}
\alias{print.pattc_ensemble}
\title{print.metalearner_ensemble}
\usage{
\method{print}{pattc_ensemble}(x, ...)

\method{print}{pattc_ensemble}(x, ...)

\method{print}{pattc_ensemble}(x, ...)

\method{print}{pattc_ensemble}(x, ...)
}
\arguments{
\item{x}{\code{pattc_deepneural} class object from \code{pattc_deepneural}}

\item{...}{additional parameter}
}
\value{
list of model results

list of model results

list of model results

list of model results
}
\description{
Print method for \code{metalearner_ensemble}

Print method for \code{metalearner_deepneural}

Print method for \code{pattc_ensemble}

Print method for \code{pattc_deepneural}
}
